+++ common/src/main/java/me/limeice/common/base/rx/RxSchedulers.java
@@ -1,7 +1,6 @@
 package me.limeice.common.base.rx;
 
 
-import io.reactivex.Observable;
 import io.reactivex.ObservableTransformer;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
